cloud-function/requirements.txt
google-auth==2.12.0
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-storage==2.5.0
pandas==1.5.2